fix: Linux wayland compilation error due xdg not found + moving xdg
